Framework_MWC_Testing/requirements.txt
selenium>=4.22.0
pytest>=8.2.0
allure-pytest>=2.13.5
pandas>=2.2.2
openpyxl>=3.1.2
requests>=2.31.0
PyYAML>=6.0.3
jsonschema>=4.26.0
pandas>=2.2.2
openpyxl>=3.1.5
